Book printing knowledge: Characteristics of Hot stamping process
Book printing knowledge: Characteristics of Hot stamping process

The hot stamping process mainly includes the following.

Characteristics:
The pattern is clear and beautiful, the color is bright and eye-catching, and it is wear-resistant. In the printed cigarette labels, the application of bronzing process accounts for more than 85%. As for hot stamping in graphic design, the China printing company will use the hot stamping process, which can play a finishing touch and highlight the design theme, especially when used for trademarks and registered names, the effect is more significant.

 Applications:
  1. Ordinary gold, laser gold hot stamping. It can be used for hot stamping on most products, including gold cards, silver cards, laser cards and glass cards, and its application is more common.

  2. Positioning holographic hot stamping. Holographic laser positioning anodized aluminum is a special process application in design, with corresponding anti-counterfeiting patterns, which can greatly improve the anti-counterfeiting ability of the product, and at the same time improve the grade of the product. Holographic laser positioning anodized aluminum has a great relationship with the bronzing temperature, bronzing pressure and vehicle speed, and even the bronzing model.

　　3. The special application of anodized aluminum in the process. Hot stamping before printing is a new idea in the cigarette label design and printing industry. The application of this process has high requirements for anodized aluminum. In addition to achieving accurate bronzing position, smooth and bright surface, balanced and uniform pressure, no blistering, no paste, etc., special attention must be paid to the edge of the bronzing pattern not to have obvious pressure The hot stamping anodized aluminum should have good adhesion on the hot stamping surface, and there should be no obvious scratches and scratches. Everyone knows that the bronzing process is actually a process in which the paper is heated and compressed. For white cards and glass cards, special attention should be paid to the protection of semi-finished products, and various unfavorable factors of paper deformation should be minimized during the production process. This will greatly help the smooth progress of the hot stamping process and the improvement of the product qualification rate.

In the application of Hot stamping process, it will encounter the problem of how to improve work efficiency in production, especially for the process improvement of large-volume products.

Direct hot stamping on the surface of UV varnish. The process itself has very high requirements for UV varnish and bronzing anodized aluminum. When coating the varnish, pay attention to controlling the amount of oil and try to ensure the relative coating of the entire batch of products. Stable, and the oil layer should be thin and flat; for the anodized aluminum used for bronzing, it is required to be high temperature resistant, have good adhesion, and have good bronzing compatibility with the varnish (the resin used in the varnish The type matches the hot melt adhesive in the bronzing anodized aluminum). In all products, the application range of this process is relatively small. In addition, the hot stamping of anodized aluminum on gold, silver ink and matt oil also has high requirements on the hot stamping surface and hot stamping anodized aluminum. If the ink particles on the hot stamping surface are large Or rough surface will directly affect the normal process of hot stamping.

If you want to custom printing China with the hot stamping process of books, please feel free to contact us, we are one of the most professional printers in China.